fix: kill switch store
